+++ src/bedrock/network/packet.h
@@ -8,6 +8,7 @@
 #include "bedrock/common_types.h"
 #include "bedrock/core/utility/binary_stream.h"
 #include "bedrock/network/minecraft_packet_ids.h"
+#include "bedrock/network/network_peer.h"
 #include "bedrock/network/serialization_mode.h"
 #include "bedrock/platform/result.h"
 
@@ -36,10 +37,23 @@ public:
     virtual void setSerializationMode(SerializationMode mode);
     [[nodiscard]] virtual std::string toString() const;
 
-    Bedrock::Result<void> readNoHeader(ReadOnlyBinaryStream &stream, const cereal::ReflectionCtx &reflection_ctx,
-                                       const SubClientId &sub_id);
+    Bedrock::Result<void> readDetour(ReadOnlyBinaryStream &stream, const cereal::ReflectionCtx &reflection_ctx);
+    void writeDetour(BinaryStream &stream, const cereal::ReflectionCtx &reflection_ctx,
+                     std::optional<SerializationMode> override_mode) const;
+
+    [[nodiscard]] SubClientId getSenderSubId() const { return sender_sub_id_; }
+    void setSenderSubId(const SubClientId sender_sub_id) { sender_sub_id_ = sender_sub_id; }
 
 private:
     virtual Bedrock::Result<void> _read(ReadOnlyBinaryStream &stream, const cereal::ReflectionCtx &reflection_ctx);
     virtual Bedrock::Result<void> _read(ReadOnlyBinaryStream &stream) = 0;
+
+    int priority_;
+    NetworkPeer::Reliability reliability_;
+    SubClientId sender_sub_id_;
+    bool is_handled_;
+    NetworkPeer::PacketRecvTimepoint recv_timepoint_;
+    const void *handler_;
+    Compressibility compressible_;
 };
+static_assert(sizeof(Packet) == 48);